backend/app/schemas/auth.py
"""Auth schemas — request/response shapes for auth endpoints."""
from pydantic import BaseModel, EmailStr


class UserCreate(BaseModel):
    email: EmailStr
    password: str  # min 8 chars — validated in endpoint


class UserLogin(BaseModel):
    email: EmailStr
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
