backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# LangGraph / LangChain
langgraph==0.2.28
langgraph-checkpoint-postgres==0.1.9
langchain-core==0.3.10
langchain-google-genai==2.0.4

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP client (for Rozee scraper)
httpx==0.27.2
beautifulsoup4==4.12.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
